Added IRLBA utilities for realized and tatami matrices. (#1)
This allows us to run IRLBA from tatami Matrices, either directly (to save
memory) or by realizing it to a temporary CSR matrix (to save time). The goal
is to ultimately use it for iterative LSI. 

We need to vendor in the CppIrlba library, along with some updates to tatami. I
don't want to vendor in Eigen but RcppEigen is not yet at 3.4.0, so instead
we'll just download the bundle in ./configure.