Create a tile matrix from the fragment file. (#3)
Pretty much as it says; runs through a fragment file and counts start/end
positions into tiles. We only use one end of the fragment to ensure that the
sum of counts across tiles is equal to the number of fragments, and to avoid
artifacts such as an overrepresentation of even counts and inflated
autocorrelations between adjacent bins.